test (resource): add ontology mock to  utils
