+++ test/data/api/v2/mockOntology.ts
@@ -1,5 +1,6 @@
 import { JsonConvert, OperationMode, ValueCheckingMode } from "json2typescript";
 import { PropertyMatchingRule } from "json2typescript/src/json2typescript/json-convert-enums";
+import { IResourceClassAndPropertyDefinitions } from "../../../../src/cache/OntologyCache";
 import { ReadOntology } from "../../../../src/models/v2/ontologies/read-ontology";
 import { ResourceClassDefinition } from "../../../../src/models/v2/ontologies/resource-class-definition";
 import { ResourcePropertyDefinition } from "../../../../src/models/v2/ontologies/resource-property-definition";
@@ -71,6 +72,64 @@ export namespace MockOntology {
         onto.dependsOnOntologies = referencedOntologies;
 
         return onto;
+
+    };
+
+    export const mockIResourceClassAndPropertyDefinitions = (resClassIri: string): IResourceClassAndPropertyDefinitions => {
+
+        const entityMock: IResourceClassAndPropertyDefinitions = {
+            classes: {},
+            properties: {}
+        };
+
+        const jsonConvert: JsonConvert = new JsonConvert(
+            OperationMode.ENABLE,
+            ValueCheckingMode.DISALLOW_NULL,
+            false,
+            PropertyMatchingRule.CASE_STRICT
+        );
+
+        const anythingOntology: any = anythingOntologyExpanded;
+        const knoraApiOntology: any = knoraApiOntologyExpanded;
+
+        const knoraApiEntities = (knoraApiOntology as { [index: string]: object[] })["@graph"];
+        const anythingEntities = (anythingOntology as { [index: string]: object[] })["@graph"];
+
+        const entities = knoraApiEntities.concat(anythingEntities);
+
+        // this.jsonConvert.operationMode = OperationMode.LOGGING;
+
+        // Convert resource classes
+        entities.filter(OntologyConversionUtils.filterResourceClassDefinitions)
+            .map(resclassJsonld => OntologyConversionUtils.convertEntity(resclassJsonld, ResourceClassDefinition, jsonConvert))
+            .filter(resclassDef => resclassDef.id === resClassIri)
+            .forEach((resClass: ResourceClassDefinition) => {
+                entityMock.classes[resClass.id] = resClass;
+            });
+
+        // properties of anything Thing
+        const props: string[]
+            = entityMock.classes[resClassIri].propertiesList.map(prop => prop.propertyIndex);
+
+        // Convert resource properties (properties pointing to Knora values)
+        entities.filter(OntologyConversionUtils.filterResourcePropertyDefinitions)
+            .map((propertyJsonld: any) => {
+                return OntologyConversionUtils.convertEntity(propertyJsonld, ResourcePropertyDefinition, jsonConvert);
+            }).filter(propertyDef => props.indexOf(propertyDef.id) !== -1)
+            .forEach((prop: ResourcePropertyDefinition) => {
+                entityMock.properties[prop.id] = prop;
+            });
+
+        // Convert system properties (properties not pointing to Knora values)
+        entities.filter(OntologyConversionUtils.filterSystemPropertyDefintions)
+            .map((propertyJsonld: any) => {
+                return OntologyConversionUtils.convertEntity(propertyJsonld, SystemPropertyDefinition, jsonConvert);
+            }).filter(propertyDef => props.indexOf(propertyDef.id) !== -1)
+            .forEach((prop: SystemPropertyDefinition) => {
+                entityMock.properties[prop.id] = prop;
+            });
+
+        return entityMock;
 
     };
